moves things into src/
